fix logic error in multiplexor
If the queue has no "in flight" records it occasionally went off
searching for such a thing anyway.  This should no longer happen.  Also
fixes workers sometimes exiting for no apparent reason (the non-apparent
reason was access to a free data structure).

+++ bwtaln.h
@@ -122,7 +122,7 @@ enum pair_phase {
 } ;
 
 typedef struct {
-    int recno ;             // to put them back into the correct order
+    long recno ;             // to put them back into the correct order
     enum pair_kind kind ;
     enum pair_phase phase ;
     bam1_t bam_rec[2] ;
